refactor: move EyesightFormValues into types/
