+++ src/types/eyesight.ts
@@ -1,0 +1,11 @@
+export type EyesightFormValues = {
+  leftEyeDegree?: string | null;
+  rightEyeDegree?: string | null;
+  leftEyePinhole?: string | null;
+  rightEyePinhole?: string | null;
+  leftAstigmatism?: string | null;
+  rightAstigmatism?: string | null;
+  leftPrescribedGlassesDegree?: string | null;
+  rightPrescribedGlassesDegree?: string | null;
+  comments?: string | null;
+};
